% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetRdConsumer}
\alias{GetRdConsumer}
\title{GetRdConsumer}
\usage{
GetRdConsumer(properties, values)
}
\arguments{
\item{properties}{a character vector indicating option properties to parameterize the RdKafka::Consumer}

\item{values}{a character vector indicating option values to parameterize the RdKafka::Consumer. Must be of same length as properties.}
}
\value{
a Rcpp::XPtr\url{RdKafka::Consumer}
}
\description{
Creates an Rcpp::XPtr\url{RdKafka::Consumer}. For more details on options \link{https://github.com/confluentinc/librdkafka/blob/master/CONFIGURATION.md}
}
