% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetRdProducer}
\alias{GetRdProducer}
\title{GetRdProducer}
\usage{
GetRdProducer(properties, values)
}
\arguments{
\item{properties}{a character vector indicating option properties to parameterize the RdKafka::Producer}

\item{values}{a character vector indicating option values to parameterize the RdKafka::Producer. Must be of same length as properties.}
}
\value{
a Rcpp::XPtr\url{RdKafka::Producer}
}
\description{
Creates an Rcpp::XPtr\url{RdKafka::Producer}. For more details on options \link{https://github.com/edenhill/librdkafka/blob/master/CONFIGURATION.md}
}
