% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KafkaConsumer.R
\name{KafkaConsumer}
\alias{KafkaConsumer}
\title{Kakfa Consumer}
\description{
A consumer is an application which
subscribes to one or more topics and processes
new messages as they arrive on that topic.

\if{html}{\out{<div class="sourceCode">}}\preformatted{         Consumers may also be instructed to process older
         messages.

         Consumers belong to "consumer groups". Consumer
         groups with more than one consumer instance can be
         used to parallelize message processing and / or
         distribute it over multiple physical machines.
}\if{html}{\out{</div>}}
}
\references{
\href{https://kafka.apache.org/documentation/#intro_consumers}{Apache Kafka docs - Consumers}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-KafkaConsumer-new}{\code{KafkaConsumer$new()}}
\item \href{#method-KafkaConsumer-subscribe}{\code{KafkaConsumer$subscribe()}}
\item \href{#method-KafkaConsumer-consume}{\code{KafkaConsumer$consume()}}
\item \href{#method-KafkaConsumer-getTopics}{\code{KafkaConsumer$getTopics()}}
\item \href{#method-KafkaConsumer-getHostPort}{\code{KafkaConsumer$getHostPort()}}
\item \href{#method-KafkaConsumer-clone}{\code{KafkaConsumer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KafkaConsumer-new"></a>}}
\if{latex}{\out{\hypertarget{method-KafkaConsumer-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KafkaConsumer$new(host, port, group_id, extra_options = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{host}}{string}

\item{\code{port}}{string}

\item{\code{group_id}}{string}

\item{\code{extra_options}}{list}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisible. Logical. \code{TRUE} if all went good.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KafkaConsumer-subscribe"></a>}}
\if{latex}{\out{\hypertarget{method-KafkaConsumer-subscribe}{}}}
\subsection{Method \code{subscribe()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KafkaConsumer$subscribe(topics)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{topics}}{string vector}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
@return invisible. Logical. \code{TRUE} if all went good.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KafkaConsumer-consume"></a>}}
\if{latex}{\out{\hypertarget{method-KafkaConsumer-consume}{}}}
\subsection{Method \code{consume()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KafkaConsumer$consume(topic, num_results = 100, timeout_ms = 1000)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{topic}}{string}

\item{\code{num_results}}{integer}

\item{\code{timeout_ms}}{integer}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisible. Logical. \code{TRUE} if all went good.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KafkaConsumer-getTopics"></a>}}
\if{latex}{\out{\hypertarget{method-KafkaConsumer-getTopics}{}}}
\subsection{Method \code{getTopics()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KafkaConsumer$getTopics()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
invisible. Logical. \code{TRUE} if all went good.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KafkaConsumer-getHostPort"></a>}}
\if{latex}{\out{\hypertarget{method-KafkaConsumer-getHostPort}{}}}
\subsection{Method \code{getHostPort()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KafkaConsumer$getHostPort()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
invisible. Logical. \code{TRUE} if all went good.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KafkaConsumer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-KafkaConsumer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KafkaConsumer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
