% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RdAssign}
\alias{RdAssign}
\title{RdAssign}
\usage{
RdAssign(consumerPtr, topic, partition)
}
\arguments{
\item{consumerPtr}{pointer. A reference to a Rcpp::XPtr\url{RdKafka::KafkaConsumer}.}

\item{topic}{string.}

\item{partition}{integer.}
}
\value{
integer. Representation of the librdkafka error code of the response to subscribe. 0 is good.
}
\description{
In process
}
