% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RdConsume}
\alias{RdConsume}
\title{RdConsume}
\usage{
RdConsume(consumerPtr, numResults, timeoutMs)
}
\arguments{
\item{consumerPtr}{pointer. A reference to a RdKafka::KafkaConsumer.}

\item{numResults}{integer. How many results should be consumed before returning. Will return early if offset is at maximum.}

\item{timeoutMs}{integer. Number of milliseconds to wait for a new message.}
}
\value{
list. With length numResults and elements topic, key and payload.
}
\description{
Consume a fixed number of results from whatever topic(s) the provided consumer is subscribed to.
}
