% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RdConsumePartition}
\alias{RdConsumePartition}
\title{RdConsumePartition}
\usage{
RdConsumePartition(consumerPtr, topic, partition, numResults, timeoutMs)
}
\arguments{
\item{consumerPtr}{pointer. A reference to a Rcpp::XPtr\url{RdKafka::KafkaConsumer}.}

\item{topic}{string.}

\item{partition}{integer.}
}
\value{
list. With length numReceived and elements topic, partition, offset, key and payload.
}
\description{
In process
}
