% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RdKafkaConsumer}
\alias{RdKafkaConsumer}
\title{RdKafkaConsumer}
\usage{
RdKafkaConsumer(properties, values)
}
\arguments{
\item{properties}{string vector. Indicating option properties to parameterize the RdKafka::Consumer.}

\item{values}{string vector. Indicating option values to parameterize the RdKafka::Consumer. Must be of same length as properties.}
}
\value{
Rcpp::XPtr\url{RdKafka::Consumer} pointer.
}
\description{
Creates an Rcpp::XPtr\url{RdKafka::Consumer}. For more details on options see \href{https://github.com/confluentinc/librdkafka/blob/master/CONFIGURATION.md}{librdkafka}.
}
