% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{KafkaConsume}
\alias{KafkaConsume}
\title{KafkaConsume}
\usage{
KafkaConsume(consumerPtr, numResults, timeoutMs)
}
\arguments{
\item{consumerPtr}{a reference to a Rcpp::XPtr\url{RdKafka::KafkaConsumer}}

\item{numResults}{how many results should be consumed before returning. Will return early if offset is at maximum}

\item{timeoutMs}{number of milliseconds to wait for a new message}
}
\value{
a list of length numResults with values list(key = key, value = value)
}
\description{
Consume a fixed number of results from whatever topic(s)
the provided consumer is subscribed to.
}
