% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KafkaConsumer.R
\name{KafkaConsumer}
\alias{KafkaConsumer}
\title{Kakfa Consumer}
\description{
A consumer is an application which subscribes to one or more topics and processes new messages as they arrive on that topic.
}
\examples{
\dontrun{
library(rdkafka)

# KafkaProducer
producer <- KafkaProducer$new(host = "localhost", port = 9092)
producer$produce(topic = "MyTest", key = "Message 1", value = "My First Message")

# KafkaConsumer
consumer <- KafkaConsumer$new(host = "localhost", port = 9092, group_id = "test",
extra_options = list(`auto.offset.reset` = "earliest"))
consumer$subscribe(topics = "MyTest")
result <- consumer$consume()
result
}
}
\references{
\href{https://kafka.apache.org/documentation/#intro_consumers}{Apache Kafka docs - Consumers}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-KafkaConsumer-new}{\code{KafkaConsumer$new()}}
\item \href{#method-KafkaConsumer-subscribe}{\code{KafkaConsumer$subscribe()}}
\item \href{#method-KafkaConsumer-assign}{\code{KafkaConsumer$assign()}}
\item \href{#method-KafkaConsumer-consume}{\code{KafkaConsumer$consume()}}
\item \href{#method-KafkaConsumer-consume_partition}{\code{KafkaConsumer$consume_partition()}}
\item \href{#method-KafkaConsumer-get_topics}{\code{KafkaConsumer$get_topics()}}
\item \href{#method-KafkaConsumer-get_brokers}{\code{KafkaConsumer$get_brokers()}}
\item \href{#method-KafkaConsumer-clone}{\code{KafkaConsumer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KafkaConsumer-new"></a>}}
\if{latex}{\out{\hypertarget{method-KafkaConsumer-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KafkaConsumer$new(brokers, group_id, extra_options = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{brokers}}{string vector. Initial list of brokers with the structure broker host or host:port.}

\item{\code{group_id}}{string. Client group id. All clients sharing the same group.id belong to the same group.}

\item{\code{extra_options}}{list. Indicating option properties and option values to parameterize the RdKafka::Consumer.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisible logical. \code{TRUE} if all went good.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KafkaConsumer-subscribe"></a>}}
\if{latex}{\out{\hypertarget{method-KafkaConsumer-subscribe}{}}}
\subsection{Method \code{subscribe()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KafkaConsumer$subscribe(topics)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{topics}}{string vector. Listing the topics to subscribe to.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisible integer. Representation of the \code{librdkafka} error code of the response to subscribe. 0 is good.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KafkaConsumer-assign"></a>}}
\if{latex}{\out{\hypertarget{method-KafkaConsumer-assign}{}}}
\subsection{Method \code{assign()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KafkaConsumer$assign(topic, partition = 0, offset = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{topic}}{string. Listing the topics to subscribe to.}

\item{\code{partition}}{integer. Indicating the partition to produce to.}

\item{\code{offset}}{integer. To do ...}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisible integer. Representation of the \code{librdkafka} error code of the response to subscribe. 0 is good.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KafkaConsumer-consume"></a>}}
\if{latex}{\out{\hypertarget{method-KafkaConsumer-consume}{}}}
\subsection{Method \code{consume()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KafkaConsumer$consume(num_results = 100, timeout_ms = 1000)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{num_results}}{integer. How many results should be consumed before returning. Will return early if offset is at maximum.}

\item{\code{timeout_ms}}{integer. Number of milliseconds to wait for a new message.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
list. Messages consumed with elements topic, key and payload.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KafkaConsumer-consume_partition"></a>}}
\if{latex}{\out{\hypertarget{method-KafkaConsumer-consume_partition}{}}}
\subsection{Method \code{consume_partition()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KafkaConsumer$consume_partition(num_results = 100, timeout_ms = 1000)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{num_results}}{integer. How many results should be consumed before returning. Will return early if offset is at maximum.}

\item{\code{timeout_ms}}{integer. Number of milliseconds to wait for a new message.}

\item{\code{topic}}{string.}

\item{\code{partition}}{integer.}

\item{\code{offset}}{integer.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
list. Messages consumed with elements topic, key and payload.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KafkaConsumer-get_topics"></a>}}
\if{latex}{\out{\hypertarget{method-KafkaConsumer-get_topics}{}}}
\subsection{Method \code{get_topics()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KafkaConsumer$get_topics()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
string vector. Listing the topics subscribed to.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KafkaConsumer-get_brokers"></a>}}
\if{latex}{\out{\hypertarget{method-KafkaConsumer-get_brokers}{}}}
\subsection{Method \code{get_brokers()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KafkaConsumer$get_brokers()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
string vector. List of brokers with the structure broker host or host:port.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KafkaConsumer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-KafkaConsumer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KafkaConsumer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
