% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KafkaProducer.R
\name{KafkaProducer}
\alias{KafkaProducer}
\title{Kakfa Producer}
\description{
A producer is an application that is responsible for publishing data to topics.
}
\references{
\href{https://kafka.apache.org/documentation/#intro_producers}{Apache Kafka docs - Producers}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-KafkaProducer-new}{\code{KafkaProducer$new()}}
\item \href{#method-KafkaProducer-produce}{\code{KafkaProducer$produce()}}
\item \href{#method-KafkaProducer-get_servers}{\code{KafkaProducer$get_servers()}}
\item \href{#method-KafkaProducer-clone}{\code{KafkaProducer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KafkaProducer-new"></a>}}
\if{latex}{\out{\hypertarget{method-KafkaProducer-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KafkaProducer$new(host, port, extra_options = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{host}}{string}

\item{\code{port}}{string}

\item{\code{extra_options}}{list}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
@return invisible. Logical. \code{TRUE} if all went good.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KafkaProducer-produce"></a>}}
\if{latex}{\out{\hypertarget{method-KafkaProducer-produce}{}}}
\subsection{Method \code{produce()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KafkaProducer$produce(topic, keys, values, partition = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{topic}}{string}

\item{\code{keys}}{string vector}

\item{\code{values}}{string vector}

\item{\code{partition}}{integer}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisible. Logical. \code{TRUE} if all went good.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KafkaProducer-get_servers"></a>}}
\if{latex}{\out{\hypertarget{method-KafkaProducer-get_servers}{}}}
\subsection{Method \code{get_servers()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KafkaProducer$get_servers()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
invisible. Logical. \code{TRUE} if all went good.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KafkaProducer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-KafkaProducer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KafkaProducer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
