% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KafkaProducer.R
\name{KafkaProducer}
\alias{KafkaProducer}
\title{Kakfa Producer}
\description{
A producer is an application that is responsible for publishing data to topics.
}
\examples{
\dontrun{
library(rdkafka)

# KafkaProducer
producer <- KafkaProducer$new()
no_sent <- producer$produce(topics = "MyTest", keys = "Message 1", payloads = "My First Message")
no_sent
}
}
\references{
\href{https://kafka.apache.org/documentation/#intro_producers}{Apache Kafka docs - Producers}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-KafkaProducer-new}{\code{KafkaProducer$new()}}
\item \href{#method-KafkaProducer-produce}{\code{KafkaProducer$produce()}}
\item \href{#method-KafkaProducer-get_brokers}{\code{KafkaProducer$get_brokers()}}
\item \href{#method-KafkaProducer-clone}{\code{KafkaProducer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KafkaProducer-new"></a>}}
\if{latex}{\out{\hypertarget{method-KafkaProducer-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KafkaProducer$new(brokers = "localhost:9092", extra_options = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{brokers}}{string vector. Initial list of brokers with the structure broker host or host:port.}

\item{\code{extra_options}}{list. Indicating option properties and option values to parameterize the RdKafka::Producer.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisible logical. \code{TRUE} if all went good.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KafkaProducer-produce"></a>}}
\if{latex}{\out{\hypertarget{method-KafkaProducer-produce}{}}}
\subsection{Method \code{produce()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KafkaProducer$produce(
  topics = "Topic",
  keys = "Key",
  payloads = "Message",
  partitions = 0L
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{topics}}{string vector. Indicating the topics to produce to. Must be of same length as keys or length equal 1.}

\item{\code{keys}}{string vector. With all the keys for the messages.}

\item{\code{payloads}}{string vector. With all the payloads for the messages. Must be of same length as keys.}

\item{\code{partitions}}{integer vector. Indicating the partitions to produce to. Must be of same length as keys or length equal 1.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisible integer. Number of messages succesfully sent.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KafkaProducer-get_brokers"></a>}}
\if{latex}{\out{\hypertarget{method-KafkaProducer-get_brokers}{}}}
\subsection{Method \code{get_brokers()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KafkaProducer$get_brokers()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
string vector. List of brokers with the structure broker host or host:port.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KafkaProducer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-KafkaProducer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KafkaProducer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
