% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RdAssign}
\alias{RdAssign}
\title{RdAssign}
\usage{
RdAssign(consumerPtr, topics, partitions, offsets)
}
\arguments{
\item{consumerPtr}{pointer. A reference to a RdKafka KafkaConsumer.}

\item{topics}{string vector. Listing the topics to subscribe to.}

\item{partitions}{integer vector.}

\item{offsets}{integer vector.}
}
\value{
integer. Representation of the librdkafka error code of the response to subscribe. 0 is good.
}
\description{
In process
}
