% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RdConsumePartition}
\alias{RdConsumePartition}
\title{RdConsumePartition}
\usage{
RdConsumePartition(consumerPtr, numResults, timeoutMs)
}
\arguments{
\item{consumerPtr}{pointer. A reference to a Rcpp::XPtr\url{RdKafka::KafkaConsumer}.}

\item{numResults}{integer. How many results should be consumed before returning. Will return early if offset is at maximum.}

\item{timeoutMs}{integer. Number of milliseconds to wait for a new message.}
}
\value{
list. With length numReceived and elements topic, partition, offset, key and payload.
}
\description{
In process
}
